# [Java.Interop.Tools.JavaCallableWrappers] use less System.Linq for CAs