--- a/src/Java.Interop.Tools.Cecil/Java.Interop.Tools.Cecil/CustomAttributeProviderRocks.cs
+++ b/src/Java.Interop.Tools.Cecil/Java.Interop.Tools.Cecil/CustomAttributeProviderRocks.cs
@@ -8,9 +8,21 @@ namespace Java.Interop.Tools.Cecil {
 
 	public static class CustomAttributeProviderRocks
 	{
+		public static bool AnyCustomAttributes (this ICustomAttributeProvider item, Type attribute) =>
+			item.AnyCustomAttributes (attribute.FullName);
+
 		public static IEnumerable<CustomAttribute> GetCustomAttributes (this ICustomAttributeProvider item, Type attribute)
 		{
 			return item.GetCustomAttributes (attribute.FullName);
+		}
+
+		public static bool AnyCustomAttributes (this ICustomAttributeProvider item, string attribute_fullname)
+		{
+			foreach (CustomAttribute custom_attribute in item.CustomAttributes) {
+				if (custom_attribute.Constructor.DeclaringType.FullName == attribute_fullname)
+					return true;
+			}
+			return false;
 		}
 
 		public static IEnumerable<CustomAttribute> GetCustomAttributes (this ICustomAttributeProvider item, string attribute_fullname)